Renamed unknown_0ABDF0 to weather
