+++ Makefile
@@ -273,6 +273,7 @@ $(call DEFINE_SECTION_FILES,audio,DEFINE_BINARY_FILES)
 $(call DEFINE_SECTION_FILES,bin,DEFINE_BINARY_FILES)
 $(call DEFINE_SECTION_FILES,fonts,DEFINE_JSON_AND_BINARY_FILES)
 $(call DEFINE_SECTION_FILES,ids,DEFINE_BINARY_FILES)
+$(call DEFINE_SECTION_FILES,levels,DEFINE_BINARY_FILES)
 $(call DEFINE_SECTION_FILES,levels/headers,DEFINE_JSON_FILES)
 $(call DEFINE_SECTION_FILES,levels/models,DEFINE_COMPRESSED_BINARY_FILES)
 $(call DEFINE_SECTION_FILES,levels/names,DEFINE_JSON_FILES)
@@ -411,6 +412,7 @@ $(eval $(call CREATE_BINARY_ASSET_TARGET,bin,Copying Bin Binary))
 $(eval $(call CREATE_BINARY_ASSET_TARGET,fonts,Copying Fonts Binary))
 $(eval $(call CREATE_JSON_ASSET_TARGET,fonts,Building Fonts))
 $(eval $(call CREATE_BINARY_ASSET_TARGET,ids,Copying IDs Binary))
+$(eval $(call CREATE_BINARY_ASSET_TARGET,levels,Copying Levels Binary))
 $(eval $(call CREATE_JSON_ASSET_TARGET,levels/headers,Building Level Header))
 $(eval $(call CREATE_COMPRESSED_ASSET_TARGET,levels/models,Compressing Level Model Binary))
 $(eval $(call CREATE_JSON_ASSET_TARGET,levels/names,Building Level Name))
