+++ include/asset_enums.h
@@ -25,7 +25,7 @@ typedef enum {
     ASSET_MISC_TABLE,
     ASSET_HUD_ELEMENT_IDS,
     ASSET_MENU_ELEMENT_IDS,
-    ASSET_UNKNOWN_19,
+    ASSET_WEATHER_PARTICLES,
     ASSET_LEVEL_OBJECT_MAPS_TABLE,
     ASSET_LEVEL_OBJECT_MAPS,
     ASSET_LEVEL_HEADERS_TABLE,
